starting to refactor for optimized search

+++ MovieFinder/Partials/MovieTitles.cs
@@ -22,7 +22,7 @@ namespace MovieFinder.Models
             }
 
             MovieTitle = movieTitlesDto.MovieTitle;
-            Year = movieTitlesDto.Year;
+            Year = (int)movieTitlesDto.Year;
         }
     }
 }
